@@ -18,7 +18,7 @@
     "@babel/plugin-proposal-object-rest-spread": "^7.12.1",
     "@babel/preset-env": "^7.12.1",
     "babel-eslint": "^10.1.0",
-    "eslint": "^7.12.1",
+    "eslint": "^7.13.0",
     "eslint-config-prettier": "^6.15.0",
     "eslint-config-standard": "^16.0.1",
     "eslint-friendly-formatter": "^4.0.1",
@@ -33,6 +33,9 @@
     "rimraf": "^3.0.2"
   },
   "dependencies": {
-    "express": "^4.17.1"
+    "apollo-server": "^2.19.0",
+    "graphql": "^15.4.0",
+    "lodash.merge": "^4.6.2",
+    "mongoose": "^5.10.14"
   }
 }
